backend/requirements.txt
# FastAPI и сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3

# Валидация данных
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Кэширование и очереди
redis==5.0.1
celery==5.3.4

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Утилиты
python-dotenv==1.0.0
httpx==0.26.0
python-dateutil==2.8.2

# ML и аналитика
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Мониторинг
prometheus-client==0.19.0

# Логирование
loguru==0.7.2

# PDF генерация
reportlab==4.0.7

